Use a generic template for all OCI Docker-compatible runtimes
